Let delegate know if CLLocationManager errors out
The only exception is if the CLLocationManager returns
an error with code kCLErrorLocationUnknown. In that case,
PPLocationFetcher pretends nothing happened and keeps
listening for updates.

As it turns out CLLocationManager doesn't stop trying to fetch
a location (hopefully with better accuracy) even if it errors out
with the code kCLErrorLocationUnknown.